+++ public/app/about/about.component.js
@@ -1,0 +1,12 @@
+import {textField, snackbar} from 'material-components-web'
+import AboutController from './about.controller'
+
+var template = require('./about.html')
+var bindings = {};
+var controller = AboutController
+
+export {
+  template,
+  bindings,
+  controller
+}