+++ public/app/authentication/local.signup.js
@@ -1,47 +1,56 @@
-var zxcvbn = require('zxcvbn');
+import {textField, linearProgress, snackbar} from 'material-components-web';
+import zxcvbn from 'zxcvbn'
 
-module.exports = {
-  template: require('./local.signup.html'),
-  bindings: {
-    strategy: '<'
+var passwordStrengthMap = {
+  0: {
+    type: 'danger',
+    text: 'Weak'
   },
-  controller: LocalSignupController
+  1: {
+    type: 'warning',
+    text: 'Fair'
+  },
+  2: {
+    type: 'info',
+    text: 'Good'
+  },
+  3: {
+    type: 'primary',
+    text: 'Strong'
+  },
+  4: {
+    type: 'success',
+    text: 'Excellent'
+  }
 };
 
-LocalSignupController.$inject = ['$scope', 'UserService'];
+class LocalSignupController {
+  constructor(UserService, $element, $timeout) {
+    this._UserService = UserService;
+    this._$element = $element;
+    this._$timeout = $timeout;
+    this.passwordStrengthScore = 0;
+  }
 
-function LocalSignupController($scope, UserService) {
-  var passwordStrengthMap = {
-    0: {
-      type: 'danger',
-      text: 'Weak'
-    },
-    1: {
-      type: 'warning',
-      text: 'Fair'
-    },
-    2: {
-      type: 'info',
-      text: 'Good'
-    },
-    3: {
-      type: 'primary',
-      text: 'Strong'
-    },
-    4: {
-      type: 'success',
-      text: 'Excellent'
-    }
-  };
-
-  this.$onInit = function() {
+  $onInit() {
     this.user = {};
   };
 
-  this.signup = function(form) {
-    var self = this;
+  $postLink() {
+    var elements = this._$element.find('.mdc-text-field');
+    for (var i = 0; i < elements.length; i++) {
+      if (i === elements.length - 1) {
+        this.passwordconfirm = new textField.MDCTextField(elements[i])
+      } else {
+        new textField.MDCTextField(elements[i]);
+      }
+    }
+    this.passwordStrengthProgress = new linearProgress.MDCLinearProgress(this._$element.find('.mdc-linear-progress')[0]);
+    this.snackbar = new snackbar.MDCSnackbar(this._$element.find('.mdc-snackbar')[0]);
+  }
 
-    form.passwordconfirm.$setValidity("nomatch", this.user.password === this.user.passwordconfirm);
+  signup(form) {
+    this.passwordconfirm.valid = this.user.password === this.user.passwordconfirm;
 
     if (!form.$valid) return;
 
@@ -53,48 +62,51 @@ function LocalSignupController($scope, UserService) {
       password: this.user.password,
       passwordconfirm: this.user.passwordconfirm
     };
+    this._UserService.signup(user, this.signupComplete.bind(this), this.signupFailed.bind(this));
+  }
 
-    // TODO throw in progress
-    var progress = function(e) {
-      if(e.lengthComputable){
-        // TODO fix me
-        // $scope.$apply(function() {
-        //   $scope.uploading = true;
-        //   $scope.uploadProgress = (e.loaded/e.total) * 100;
-        // });
-      }
-    };
+  signupComplete(data) {
+    this._$timeout(() => {
+      this.user = {};
+      this.passwordStrengthScore = 0;
+      this.onSuccess();
+      this.showStatusMessage("Success", "Account created, contact an administrator to activate your account.");
+    });
+  }
 
-    var complete = function() {
-      $scope.$apply(function() {
-        form.$submitted = false;
-        self.user = {};
-        self.passwordStrengthScore = 0;
-        self.showStatusMessage("Success", "Account created, contact an administrator to activate your account.", "alert-success");
-      });
-    };
+  signupFailed(data) {
+    this._$timeout(() => {
+      this.showStatusMessage("There was a problem creating your account", data.responseText);
+    });
+  }
 
-    var failed = function(data) {
-      $scope.$apply(function() {
-        self.showStatusMessage("There was a problem creating your account", data.responseText, "alert-danger");
-      });
-    };
-
-    UserService.signup(user, complete, failed, progress);
-  };
-
-  this.passwordStrengthScore = 0;
-  this.onPasswordChange = function() {
+  onPasswordChange() {
     var score = this.user.password && this.user.password.length ? zxcvbn(this.user.password, [this.user.username, this.user.displayName, this.user.email]).score : 0;
     this.passwordStrengthScore = score + 1;
     this.passwordStrengthType = passwordStrengthMap[score].type;
     this.passwordStrength = passwordStrengthMap[score].text;
+    this.passwordStrengthProgress.progress = this.passwordStrengthScore/5;
   };
 
-  this.showStatusMessage = function (title, message, statusLevel) {
+  showStatusMessage(title, message) {
     this.statusTitle = title;
     this.statusMessage = message;
-    this.statusLevel = statusLevel;
-    this.showStatus = true;
+    this.snackbar.open();
   };
+}
+
+LocalSignupController.$inject = ['UserService', '$element', '$timeout'];
+
+var template = require('./local.signup.html');
+var bindings = {
+  strategy: '<',
+  onCancel: '&',
+  onSuccess: '&'
+};
+var controller = LocalSignupController;
+
+export {
+  template,
+  bindings,
+  controller
 }