+++ public/app/mage/index.js
@@ -1,6 +1,7 @@
 var angular = require('angular');
 
 angular.module('mage')
+  .component('mageInfo', require('./mage-info.component.js'))
   .directive('banner', require('./banner.directive'))
   .directive('colorPicker', require('./color.picker.directive'))
   .directive('equals', require('./equals.directive'))