+++ public/img/index.js
@@ -1,9 +1,17 @@
 require('./favicon.ico');
 require('./default-icon.png');
 require('./google_play_store_btn@2x.png');
+require('./google-play-badge.png');
 require('./ios_app_store_btn@2x.png');
+require('./ios_app_store_white.svg');
+require('./ios_app_store_black.svg');
 require('./missing_marker.png');
 require('./missing_photo.png');
 require('./login-gov-badge.png');
 require('./baseline-mood-24px.svg');
 require('./baseline-account_circle-24px.svg');
+require('./1.jpg');
+require('./2.jpg');
+require('./3.jpg');
+require('./4.jpg');
+require('./bluewand.png');